Continued working on the ability system
The ability system now follows a factory pattern, which makes it rather
light weight to add new abilities to the game and gathers them mostly
locally in a global import. There is still the ability to define their
action within the game scope, allowing them to be more complex.

For this the player state global import was extended to provide a more
controlled interface to it's data structures. These allows the game to
listen for changes and adapet to them.